+++ packages/eslint-config/index.js
@@ -10,11 +10,22 @@ module.exports = {
   plugins: ["import-helpers", "perfectionist", "@typescript-eslint"],
   overrides: [
     {
+      env: {
+        node: true,
+      },
       files: ["*.ts", "*.tsx"],
-      extends: ["plugin:@typescript-eslint/recommended"],
+      extends: [
+        "plugin:@typescript-eslint/recommended",
+        // 'plugin:@typescript-eslint/recommended-requiring-type-checking',
+      ],
+      parserOptions: {
+        project: ["./tsconfig.json"],
+      },
     },
   ],
   rules: {
+    "@typescript-eslint/await-thenable": "error",
+
     "perfectionist/sort-object-types": [
       "error",
       {